fix: ghmusic lua script fix

+++ ghmusic.lua
@@ -2,7 +2,6 @@ local args = {...}
 
 if #args < 1 then
   print("Uzycie: ghmusic <folder> [numer]")
-  print("Przyklad: ghmusic bijmenela 1")
   return
 end
 
@@ -10,8 +9,15 @@ local folder = args[1]
 local num = tonumber(args[2]) or 1
 
 local base = "https://raw.githubusercontent.com/IXANOR/computercraftmusictest/main/"
-local url = base .. folder .. "/" .. tostring(num) .. ".dfpwm"
+local url = base .. folder .. "/" .. num .. ".dfpwm"
 local tmp = "temp.dfpwm"
+
+local drive = peripheral.find("tape_drive")
+
+if not drive then
+  print("Nie znaleziono napedu kasety.")
+  return
+end
 
 if fs.exists(tmp) then
   fs.delete(tmp)
@@ -23,15 +29,25 @@ print(url)
 shell.run("wget", url, tmp)
 
 if not fs.exists(tmp) then
-  print("Plik nie zostal pobrany.")
+  print("Nie udalo sie pobrac pliku.")
   return
 end
 
-print("Zapisuje na kasete...")
-shell.run("tape", "write", tmp)
+print("Nagrywam kasete...")
 
-if fs.exists(tmp) then
-  fs.delete(tmp)
+local file = fs.open(tmp, "rb")
+
+drive.stop()
+drive.seek(-drive.getSize())
+
+while true do
+  local chunk = file.read(8192)
+  if not chunk then break end
+  drive.write(chunk)
 end
+
+file.close()
+
+fs.delete(tmp)
 
 print("Gotowe.")
